fix: resolve all compiler warnings in C++ codebase
Fixed all compiler warnings across the plugin codebase:

**Unused Parameter Warnings**:
- Added comment-style suppressions for intentionally unused parameters
- Fixed in OSNotification, ErrorTracking, WebSocketClient, StreamChatClient,
  AudioCapture, MIDICapture, Microphone, ChordSequenceDetector

**Sign Conversion Warnings**:
- Added explicit static_cast<> conversions for int->size_t conversions
- Fixed in HttpErrorHandler, PerformanceMonitor, WebSocketClient,
  StreamChatClient, ErrorTracking, LogSink, OperationalTransform

**Shadow Warnings**:
- Renamed local variables to avoid shadowing class members
- Fixed in StreamChatClient (renamed token/apiKey/userId)
- Fixed in PerformanceMonitor (renamed name to metricName)

**Returning Reference to Temporary**:
- Changed DraftStore::getDrafts() return type from const ref to value

**Nullable to Non-nullable Conversion**:
- Added null checks in OSNotification_mac.mm before NSString assignments

All fixes maintain C++26 compatibility and don't change functionality.

+++ plugin/src/util/HttpErrorHandler.h
@@ -96,7 +96,7 @@ public:
         std::lock_guard<std::mutex> lock(errorMutex);
         std::vector<HttpError> result;
 
-        int startIdx = juce::jmax(0, static_cast<int>(errorHistory.size()) - count);
+        size_t startIdx = static_cast<size_t>(juce::jmax(0, static_cast<int>(errorHistory.size()) - count));
         for (size_t i = startIdx; i < errorHistory.size(); ++i)
             result.push_back(errorHistory[i]);
